ip2gid: Move server and client code to separate files.
Move server code to server.c and client code to client.c, respectively.
Add ip2gid.c/h for common code.
